requirement.txt
numpy 
pillow
scipy
opencv-python
scikit-image
tensorflow
imgaug
matplotlib
labelme2coco
torch==1.6.0